update: routes & controllers

+++ backend/controllers/controllersBooks.js
@@ -1,24 +1,24 @@
 const Book = require('../models/book.js');
+const fs = require('fs');
 
 exports.createBook = (req, res) => {
-    const fullLink = req.body.fullLink;
     const bookObject = JSON.parse(req.body.book);
     delete bookObject._id;
     delete bookObject._userId;
     const book = new Book({
         ...bookObject,
        userId: req.auth.userId,
-       imageUrl: `${fullLink}`
+       imageUrl: req.fullLink
     });
     book.save()
     .then(() => res.status(201).json({ message: 'Livre enregistré !'}))
     .catch(error => res.status(400).json({ error }));
 };
 
-exports.modifyBook = ('/:id', (req, res, next) => {
+exports.modifyBook = (req, res, next) => {
     const bookObject = req.file ? {
         ...JSON.parse(req.body.book),
-        imageUrl: `${fullLink}`
+        imageUrl: req.fullLink
     } : { ...req.body };
     
     delete bookObject._userId;
@@ -32,15 +32,28 @@ exports.modifyBook = ('/:id', (req, res, next) => {
                .catch(error => res.status(401).json({ error }));
            }
        })
-  });
+  };
 
-exports.deleteBook = ('/:id', (req, res, next) => {
-Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Livre supprimé !'}))
-    .catch(error => res.status(400).json({ error }));
-});
+  exports.deleteBook = (req, res, next) => {
+    Book.findOne({ _id: req.params.id})
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
 
-exports.getOneBook = ('/:id', (req, res, next) => {
+exports.getOneBook = ('/:id', (req, res) => {
     Book.findOne({
     _id: req.params.id
     })
@@ -48,8 +61,48 @@ exports.getOneBook = ('/:id', (req, res, next) => {
     .catch(error => res.status(400).json({ error }));
   });
 
-exports.getAllBooks = ('/' + '', (req, res, next) => {
+exports.getAllBooks = ('/' + '', (req, res) => {
     Book.find()
     .then((books) => {res.status(200).json(books);})
     .catch((error) => {res.status(400).json({error: error});});
 });
+
+
+exports.createRating = (req, res) => {
+    const newRating = {
+      userId: req.auth.userId,
+      grade: req.body.rating,
+    };
+  
+    Book.findOne({ _id: req.params.id })
+      .then((book) => {
+        if (!book) {
+          return res.status(404).json({ message: 'Book not found' });
+        }
+  
+        const existingRatingIndex = book.ratings.findIndex(rating => rating.userId === req.auth.userId);
+  
+        if (existingRatingIndex !== -1) {
+          book.ratings[existingRatingIndex].grade = newRating.grade;
+        } else {
+          book.ratings.push(newRating);
+        }
+  
+        const totalRatings = book.ratings.reduce((sum, rating) => sum + rating.grade, 0);
+        const newAverage = totalRatings / (book.ratings.length + 1);
+        book.averageRating = newAverage;
+
+        book.save()
+          .then(() => res.status(200).json({ message: 'Rating added/updated successfully!' }))
+          .catch(error => res.status(400).json({ error }));
+      })
+      .catch(error => res.status(400).json({ error }));
+  }
+
+exports.getBestRated = (req, res) => {
+    Book.find()
+      .sort({ averageRating: -1 })
+      .limit(3)
+      .then(books => res.status(200).json(books))
+      .catch(error => res.status(400).json({ error }));
+}
